server/pyproject.toml
[tool.poetry]
name = "pos-server"
version = "0.1.0"
description = "Point of Sale Server - FastAPI backend with MQTT sync"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
paho-mqtt = "^1.6.1"
python-dotenv = "^1.0.0"
alembic = "^1.12.1"
cloudinary = "^1.36.0"
python-multipart = "^0.0.6"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"
debugpy = "^1.8.20"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

